requirements.txt
flask==2.0.1
torch>=1.7.1
torchvision>=0.8.2
numpy>=1.19.2
pillow>=8.0.1
python-dotenv>=0.19.0
flask-wtf>=0.15.1
werkzeug==2.0.3
ninja>=1.10.0
tqdm>=4.62.3